Refactor stock list layout and add "Add" button
This commit refactors the layout of the stock list page by removing the unnecessary Center widget and adding an Expanded widget to make the stock list take up the available space. Additionally, a "Add" button is added at the bottom of the page, allowing users to navigate to the stock add page.

Closes #1006